Move confirmation modal to common components
